Move SqlClient fix for .NET 6+ compatibility into SqlClient-only file (#1)
