+++ src/System.Data.SqlClient/src/System.Data.SqlClient.csproj
@@ -480,6 +480,7 @@
     <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\GssSafeHandles.cs">
       <Link>Common\Microsoft\Win32\SafeHandles\GssSafeHandles.cs</Link>
     </Compile>
+    <Compile Include="Common\Interop\Unix\System.Net.Security.Native\Interop.NetSecurityNative.Extended.cs" />
     <Compile Include="$(CommonPath)\Interop\Unix\System.Net.Security.Native\Interop.NetSecurityNative.cs">
       <Link>Common\Interop\Unix\System.Net.Security.Native\Interop.NetSecurityNative.cs</Link>
     </Compile>
